uttu/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.5
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.42.0
django-audiofield==0.10.0
django-categories==1.8
django-category==2.0.1
django-celery==3.3.1
django-cropperjs==0.0.5
django-embed-video==1.3.3
django-js-asset==1.2.2
django-mptt==0.11.0
django-object-tools==2.0.0
django-rest-framework==0.1.0
django-tables2==2.3.1
djangorestframework==3.11.1
html5lib==1.1
idna==2.10
kombu==3.0.37
oauthlib==3.1.0
Pillow==7.1.1
PyPDF2==1.26.0
python3-openid==3.2.0
pytz==2019.3
reportlab==3.5.49
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
unicode-slugify==0.1.3
Unidecode==1.1.1
urllib3==1.25.10
vine==1.3.0
webencodings==0.5.1
xhtml2pdf==0.2.4
